+++ NEWS.md
@@ -1,5 +1,8 @@
 # box (devel)
 
+* Fix: Attach names starting with a dot (#186)
+
+
 # box 1.0.0
 
 Complete rewrite; see the [migration